+++ backend-core/src/main/java/com/promptpicture/backend/core/payment_method/adapter/PaymentMethodRepositoryAdapter.java
@@ -1,0 +1,10 @@
+package com.promptpicture.backend.core.payment_method.adapter;
+
+import com.promptpicture.backend.core.payment_method.domain.PaymentMethod;
+
+import java.util.List;
+
+public interface PaymentMethodRepositoryAdapter {
+
+    List<PaymentMethod> getListOfPaymentMethod();
+}