+++ changelog.md
@@ -14,6 +14,7 @@
 - aitest-5 - add job for removing prompt older than 12 hours
 - aitest-37 - add cart view and the ability to remove items
 - aitest-38 - add order_details, order_item, invoice table
+- aitest-41- add payment method table
 
 ### Fix